Unittesting positional arguments verification. (#214)
Related to that several minor changes:
rootCmd configuration moved to init() function. main only rootCmd.Execute() it.
This is the normal way to setup rootCmd, and it is needed, so when unit-testing the code inside init() is executed.
Some functions simplified to be just a function instead of a function that returns a function (exposeTargetArgs, bindArgs, etc).