Fix Unspecified Vanflow SITE Attributes (#1893)
Previously the vanflow SITE record was using an unspecified attributes
to denote the site version and policy status. This breaking change moves
the SITE Version attribute to the specified buildVersion attribute (code
point 32) in the spec and removes the now-deprecated Policy attribute.
Versions running the old unspecified version attribute (2.0.0-previews 1
and 2) will not observe the site version from sites with the corrected
attribute and vice versa. The Policy attribute has been unused since
1.x.

Signed-off-by: Christian Kruse <christian@c-kruse.com>