Use real files as targets for binaries, with dependencies (#1973)
This adds a macro using "go list" to determine the full list of package
dependencies for a given Go binary, and then builds the corresponding
list of files as prerequisites. Replacing the phony targets with
targets matching the built programs then allows Make to correctly
rebuild binaries only when necessary, saving time when regularly
rebuilding the project locally.

This requires changing the phony generate-doc target, renamed to
generate-docs; as far as I can tell it isn't used outside the project
so this shouldn't have a drastic impact.

The phony build targets are preserved for backward compatibility.

Signed-off-by: Stephen Kitt <skitt@redhat.com>