Marks test helper functions with t.Helper (#2021)
Marks several test utilities with the testing.T Helper function so that
when assertions made within that helper fail, the go testing framework
will report line number info for the calling test case instead of the
shared helper function. This should make it quicker for us to identify
the source of the test failure.

Signed-off-by: Christian Kruse <christian@c-kruse.com>