Include go vet in CI test job (#2125)
Go test runs a subset of vet's checkers, and with the large volume of tests
in our codebase (especially with verbose output set) it can be difficult
to identify the source of a test failure when it comes from a vet
violation. This adds an explicit step to run `go vet` before the tests
run.

Signed-off-by: Christian Kruse <christian@c-kruse.com>