Fixes flakey TestNodeWatcher test (#2168)
TestNodeWatcher seemed to be failing due to a race between the testing
goroutine exercising the fake k8s client (creating and deleting
resources) and the goroutine of the informer syncronizing its cache and
invoking handlers/callbacks based on the current state of a resources.
This change introduces blocking in the test goroutine - expecting
callbacks for the initial resource state to run before continuing to
update resources.

Signed-off-by: Christian Kruse <christian@c-kruse.com>